fix: address review comments for Hytte-goir
- Extract formatRelative to shared utils.ts to eliminate duplication between ChatView and ConversationList
- Memoize Intl.RelativeTimeFormat per i18n.language in both components to avoid re-creating it on every render
- Fix Composer to count Unicode code points ([...str].length) instead of UTF-16 code units (str.length) to correctly mirror the backend rune-count limit
- Remove maxLength from the textarea so the submit-time tooLong error is reachable for astral-char inputs
